check approval state and balance for akka estimateGas
