+++ apps/web/src/views/Swap/AkkaSwap/hooks/useAkkaRouterApi.ts
@@ -78,46 +78,9 @@ export const useAkkaRouterRoute = (token0: Currency, token1: Currency, amount: C
 
 // Call both apis route and args together in the same time
 export const useAkkaRouterRouteWithArgs = (token0: Currency, token1: Currency, amount: CurrencyAmount<Currency>, slippage = 0.1) => {
-  // isAkkaSwapMode checks if this is akka router form or not from redux
-  const [isAkkaSwapMode, toggleSetAkkaMode, toggleSetAkkaModeToFalse, toggleSetAkkaModeToTrue] =
-    useIsAkkaSwapModeStatus()
-  // isAkkaContractSwapMode checks if this is akka router form or not from redux
-  const [isAkkaContractSwapMode, toggleSetAkkaContractMode, toggleSetAkkaContractModeToFalse, toggleSetAkkaContractModeToTrue] =
-    useIsAkkaContractSwapModeActive()
-
   const route = useAkkaRouterRoute(token0, token1, amount, slippage)
   const args = useAkkaRouterArgs(token0, token1, amount, slippage)
-  const akkaContract = useAkkaRouterContract()
-  const { account, isConnected } = useWeb3React()
-  const methodName = 'multiPathSwap'
-  const {
-    independentField,
-    typedValue,
-    [Field.INPUT]: { currencyId: inputCurrencyId },
-    [Field.OUTPUT]: { currencyId: outputCurrencyId },
-  } = useSwapState()
-  const { chainId } = useActiveChainId()
-  if (isConnected) {
-    akkaContract.estimateGas[methodName](
-      args?.data?.amountIn,
-      args?.data?.amountOutMin,
-      args?.data?.data,
-      [],
-      [],
-      account
-      , {
-        value: inputCurrencyId === NATIVE[chainId].symbol ? args?.data?.amountIn : '0',
-      })
-      .then(() => {
-        toggleSetAkkaContractModeToTrue()
-      })
-      .catch(() => {
-        toggleSetAkkaContractModeToFalse()
-      })
-  }
-  else {
-    toggleSetAkkaContractModeToTrue()
-  }
+  
   return {
     route,
     args,
